refactor(air): adapt lookup module to 2856 unified Challenges + MainCols
Post-rebase Phase 3 of the lookup-module adaptation:

- Replace LookupChallenges struct with a type alias for trace::Challenges (the
  unified 2856 challenge type with fixed-size bus_prefix / beta_powers arrays).
- Migrate LookupMessage trait to use trace::Challenges in place of the dropped
  LookupChallenges. ConstraintLookupBuilder / ProverLookupBuilder / DualBuilder
  drop the max_message_width / num_bus_ids constructor args (sizes are now
  fixed at compile time via NUM_BUS_TYPES / MAX_MESSAGE_WIDTH).
- bus_prefix() trait method takes usize instead of u16; lookup::bus_id is now
  a thin re-export shim over trace::bus_types.
- Rewrite lookup/buses/{block_stack,block_hash_and_op_group,chiplet_requests,
  range_logcap,wiring}.rs against the typed MainCols struct: replace
  MainTraceRow indexed access with DecoderCols / StackCols / RangeCols /
  SystemCols field accessors, switch OpFlags::new from the
  ExprDecoderAccess wrapper to the (decoder, stack, decoder_next) signature
  (which folds end_next/repeat_next/halt_next into a single struct), and
  recover the ACE chiplet flag with the new virtual s0 = 1 - s_ctrl - s_perm
  formula.
- Add bus parameters to every Challenges::encode call in logup_msg.rs (each
  inherent Msg::encode now passes its bus type from trace::bus_types).
- Stub lookup/buses/{chiplet_responses,hash_kernel}.rs with no-op emit_*
  functions: the original implementations referenced the pre-2856 hasher
  chiplet (32-row P_CYCLE_ROW_0/31, hasher::flags::f_mv etc.) and need to be
  rewritten against the new hasher_control + permutation split + 16-row
  HasherPeriodicCols. Originals are preserved on adr1anh/bus.pre-rebase-backup.
- Delete dead constraints/logup.rs (the legacy RationalSet / Batch / Column
  algebra is no longer referenced; superseded by lookup module).
- Disable constraints/degree_audit.rs tests with cfg(any()) until they're
  ported to MainCols / OpFlags::new(...) / Challenges::encode(bus, [...]).

The two MidenLookupAir canary tests pass: the degree budget (with the two
stubbed columns counted) and the cached-encoding equivalence check.

+++ air/src/constraints/lookup/builder.rs
@@ -439,7 +439,7 @@ pub trait EncodedLookupGroup: LookupGroup {
     /// [`LookupAir::num_bus_ids`](super::LookupAir::num_bus_ids)
     /// contract was violated. Authors should ensure every bus ID they
     /// emit satisfies `bus_id < num_bus_ids`.
-    fn bus_prefix(&self, bus_id: u16) -> Self::ExprEF;
+    fn bus_prefix(&self, bus_id: usize) -> Self::ExprEF;
 
     /// Add a flag-gated interaction whose denominator is already an
     /// extension-field expression.
