+++ air/src/constraints/lookup/prover.rs
@@ -366,8 +366,8 @@ where
         &self.challenges.beta_powers[..]
     }
 
-    fn bus_prefix(&self, bus_id: u16) -> Self::ExprEF {
-        self.challenges.bus_prefix[bus_id as usize]
+    fn bus_prefix(&self, bus_id: usize) -> Self::ExprEF {
+        self.challenges.bus_prefix[bus_id]
     }
 
     fn insert_encoded(&mut self, flag: F, multiplicity: F, encoded: impl FnOnce() -> EF) {
